changed file structure, moved callbacks into corresponding files and created a model object in config
