+++ config.py
@@ -1,0 +1,7 @@
+from pharmpy.modeling import create_basic_pk_model, convert_model
+
+model = create_basic_pk_model('iv')
+model = convert_model(model, "nonmem")
+
+def make_label_value(key, value):
+            return {"label": key, "value":value}
